+++ test/src/base/hashTest.c
@@ -9,7 +9,8 @@
 
 static uintptr_t hash_same_fn(void *key);
 
-static unsigned int hash_str(void *key);
+static uintptr_t hash_str(void *key);
+
 static unsigned int str_equ(void *key1, void *key2);
 
 static d_hash_t hash = NULL;
@@ -183,7 +184,7 @@ static uintptr_t hash_same_fn(void *key) {
 	return (uintptr_t) key;
 }
 
-static unsigned int hash_str(void *key) {
+static uintptr_t hash_str(void *key) {
 	char *str = (void *) key;
 	unsigned i = 0;
